Add support for building for 64-bit Windows, currently only the interpreted version.  Update the project files for Visual Studio 2010 and fix various issues arising from int and unsigned being 32-bit rather than 64-bit.  Fix underscore on exported symbol in pecoffexport.cpp.
git-svn-id: svn://svn.code.sf.net/p/polyml/code/trunk/polyml@1509 ae7d391e-3f74-4a2b-a0db-a8776840fd2a

+++ libpolyml/exporter.cpp
@@ -607,7 +607,7 @@ ExportStringTable::~ExportStringTable()
 // Add a string to the string table, growing it if necessary.
 unsigned long ExportStringTable::makeEntry(const char *str)
 {
-    int len = strlen(str);
+    unsigned len = (unsigned)strlen(str);
     unsigned long entry = stringSize;
     if (stringSize + len + 1 > stringAvailable)
     {
